fixes traversal
