+++ graph-extension/src/main/java/org/jnosql/artemis/graph/VertexTraversal.java
@@ -27,105 +27,8 @@ import static java.util.Objects.requireNonNull;
  * The Graph Traversal that maps {@link org.apache.tinkerpop.gremlin.structure.Vertex}.
  * This Traversal is lazy, in other words, that just run after the
  */
-public interface VertexTraversal {
+public interface VertexTraversal extends VertexConditionTraversal {
 
-
-    /**
-     * Adds a equals condition to a query
-     *
-     * @param propertyKey the key
-     * @param value       the value to the condition
-     * @return a {@link VertexTraversal} with the new condition
-     * @throws NullPointerException when either key or value are null
-     */
-    VertexTraversal has(String propertyKey, Object value) throws NullPointerException;
-
-    /**
-     * Adds a equals condition to a query
-     *
-     * @param propertyKey the key
-     * @param predicate   the predicate condition
-     * @return a {@link VertexTraversal} with the new condition
-     * @throws NullPointerException when either key or predicate condition are null
-     */
-    VertexTraversal has(String propertyKey, P<?> predicate) throws NullPointerException;
-
-    /**
-     * Adds a equals condition to a query
-     *
-     * @param propertyKey the key
-     * @param value       the value to the condition
-     * @return a {@link VertexTraversal} with the new condition
-     * @throws NullPointerException when either key or value are null
-     */
-    default VertexTraversal has(Supplier<String> propertyKey, Object value) throws NullPointerException{
-        requireNonNull(propertyKey, "the supplier is required");
-        return has(propertyKey.get(), value);
-    }
-
-    /**
-     * Adds a equals condition to a query
-     *
-     * @param propertyKey the key
-     * @param predicate   the predicate condition
-     * @return a {@link VertexTraversal} with the new condition
-     * @throws NullPointerException when either key or predicate condition are null
-     */
-    default VertexTraversal has(Supplier<String> propertyKey, P<?> predicate) throws NullPointerException{
-        requireNonNull(propertyKey, "the supplier is required");
-        return has(propertyKey.get(), predicate);
-    }
-
-    /**
-     * Adds a equals condition to a query
-     *
-     * @param accessor the key
-     * @param value    the value to the condition
-     * @return a {@link VertexTraversal} with the new condition
-     * @throws NullPointerException when either key or value are null
-     */
-    VertexTraversal has(T accessor, Object value) throws NullPointerException;
-
-    /**
-     * Adds a equals condition to a query
-     *
-     * @param accessor  the key
-     * @param predicate the predicate condition
-     * @return a {@link VertexTraversal} with the new condition
-     * @throws NullPointerException when either key or value are null
-     */
-    VertexTraversal has(T accessor, P<?> predicate) throws NullPointerException;
-
-
-    /**
-     * Defines Vertex has not a property
-     *
-     * @param propertyKey the property key
-     * @return a {@link VertexTraversal} with the new condition
-     * @throws NullPointerException when propertyKey is null
-     */
-    VertexTraversal hasNot(String propertyKey) throws NullPointerException;
-
-    /**
-     * Defines Vertex has not a property
-     *
-     * @param propertyKey the property key
-     * @return a {@link VertexTraversal} with the new condition
-     * @throws NullPointerException when propertyKey is null
-     */
-    default VertexTraversal hasNot(Supplier<String> propertyKey) throws NullPointerException{
-        requireNonNull(propertyKey, "the supplier is required");
-        return hasNot(propertyKey.get());
-    }
-
-    /**
-     * Map the {@link VertexTraversal} to its outgoing adjacent vertices given the edge labels.
-     *
-     * @param labels the edge labels to traverse
-     * @return a {@link VertexTraversal} with the new condition
-     * @throws NullPointerException when has any null element
-     */
-    VertexTraversal out(String... labels) throws NullPointerException;
 
     /**
      * Map the {@link EdgeTraversal} to its outgoing incident edges given the edge labels.
@@ -136,14 +39,6 @@ public interface VertexTraversal {
      */
     EdgeTraversal outE(String... edgeLabels) throws NullPointerException;
 
-    /**
-     * Map the {@link VertexTraversal} to its adjacent vertices given the edge labels.
-     *
-     * @param labels the edge labels to traverse
-     * @return a {@link VertexTraversal} with the new condition
-     * @throws NullPointerException when has any null element
-     */
-    VertexTraversal in(String... labels) throws NullPointerException;
 
     /**
      * Map the {@link EdgeTraversal} to its incoming incident edges given the edge labels.
@@ -153,15 +48,6 @@ public interface VertexTraversal {
      * @throws NullPointerException when has any null element
      */
     EdgeTraversal inE(String... edgeLabels) throws NullPointerException;
-
-    /**
-     * Map the {@link VertexTraversal} to its incoming adjacent vertices given the edge labels.
-     *
-     * @param labels the edge labels to traverse
-     * @return a {@link VertexTraversal} with the new condition
-     * @throws NullPointerException when has any null element
-     */
-    VertexTraversal both(String... labels) throws NullPointerException;
 
     /**
      * Map the {@link EdgeTraversal} to its either incoming or outgoing incident edges given the edge labels.
@@ -175,21 +61,11 @@ public interface VertexTraversal {
 
     /**
      * Starts the loop traversal graph
+     *
      * @return a {@link VertexRepeatTraversal}
      */
     VertexRepeatTraversal repeat();
 
-    /**
-     * Map the {@link VertexTraversal} to its outgoing adjacent vertices given the edge labels.
-     *
-     * @param label the edge labels to traverse
-     * @return a {@link VertexTraversal} with the new condition
-     * @throws NullPointerException when has any null element
-     */
-    default VertexTraversal out(Supplier<String> label) throws NullPointerException{
-        requireNonNull(label, "the supplier is required");
-        return out(label.get());
-    }
 
     /**
      * Map the {@link EdgeTraversal} to its outgoing incident edges given the edge labels.
@@ -198,22 +74,11 @@ public interface VertexTraversal {
      * @return a {@link EdgeTraversal} with the new condition
      * @throws NullPointerException when has any null element
      */
-    default EdgeTraversal outE(Supplier<String> label) throws NullPointerException{
+    default EdgeTraversal outE(Supplier<String> label) throws NullPointerException {
         requireNonNull(label, "the supplier is required");
         return outE(label.get());
     }
 
-    /**
-     * Map the {@link VertexTraversal} to its adjacent vertices given the edge labels.
-     *
-     * @param label the edge labels to traverse
-     * @return a {@link VertexTraversal} with the new condition
-     * @throws NullPointerException when has any null element
-     */
-    default VertexTraversal in(Supplier<String> label) throws NullPointerException{
-        requireNonNull(label, "the supplier is required");
-        return in(label.get());
-    }
 
     /**
      * Map the {@link EdgeTraversal} to its incoming incident edges given the edge labels.
@@ -222,22 +87,11 @@ public interface VertexTraversal {
      * @return a {@link EdgeTraversal} with the new condition
      * @throws NullPointerException when has any null element
      */
-    default EdgeTraversal inE(Supplier<String> label) throws NullPointerException{
+    default EdgeTraversal inE(Supplier<String> label) throws NullPointerException {
         requireNonNull(label, "the supplier is required");
         return inE(label.get());
     }
 
-    /**
-     * Map the {@link VertexTraversal} to its incoming adjacent vertices given the edge labels.
-     *
-     * @param label the edge labels to traverse
-     * @return a {@link VertexTraversal} with the new condition
-     * @throws NullPointerException when has any null element
-     */
-    default VertexTraversal both(Supplier<String> label) throws NullPointerException{
-        requireNonNull(label, "the supplier is required");
-        return both(label.get());
-    }
 
     /**
      * Map the {@link EdgeTraversal} to its either incoming or outgoing incident edges given the edge labels.
@@ -246,7 +100,7 @@ public interface VertexTraversal {
      * @return a {@link EdgeTraversal} with the new condition
      * @throws NullPointerException when has any null element
      */
-    default EdgeTraversal bothE(Supplier<String> label) throws NullPointerException{
+    default EdgeTraversal bothE(Supplier<String> label) throws NullPointerException {
         requireNonNull(label, "the supplier is required");
         return bothE(label.get());
     }
@@ -260,27 +114,6 @@ public interface VertexTraversal {
      * @return a {@link VertexTraversal} with the limit
      */
     VertexTraversal limit(long limit);
-
-    /**
-     * Defines Vertex as label condition
-     *
-     * @param label the labels in the condition
-     * @return a {@link VertexTraversal} with the new condition
-     * @throws NullPointerException when has any null element
-     */
-    VertexTraversal hasLabel(String label) throws NullPointerException;
-
-    /**
-     * Defines Vertex as label condition
-     *
-     * @param label the labels in the condition
-     * @return a {@link VertexTraversal} with the new condition
-     * @throws NullPointerException when has any null element
-     */
-    default VertexTraversal hasLabel(Supplier<String> label) throws NullPointerException{
-        requireNonNull(label, "the supplier is required");
-        return hasNot(label.get());
-    }
 
 
     /**
